add discount usage everywhere
